Separates fire distribution into multiple files
