+++ CoreSystems.csproj
@@ -362,7 +362,10 @@
     <Compile Include="Data\Scripts\CoreSystems\Comms\Space.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Compile>
-    <Compile Include="Data\Scripts\CoreSystems\Support\FireDistribution.cs" />
+    <Compile Include="Data\Scripts\CoreSystems\Support\FireDistribution\AdvancedFireDistributionSystem.cs" />
+    <Compile Include="Data\Scripts\CoreSystems\Support\FireDistribution\FireDistribution.cs" />
+    <Compile Include="Data\Scripts\CoreSystems\Support\FireDistribution\Implementation\AdvancedClosestFireDistributionSystem.cs" />
+    <Compile Include="Data\Scripts\CoreSystems\Support\FireDistribution\Implementation\AdvancedScreeningFireDistributionSystem.cs" />
     <Compile Include="Data\Scripts\CoreSystems\Support\Localization.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Compile>
