+++ Data/Scripts/CoreSystems/EntityComp/Controls/Weapon/WeaponUi.cs
@@ -10,6 +10,7 @@ using VRage.ModAPI;
 using VRage.Utils;
 using VRageMath;
 using WeaponCore.Data.Scripts.CoreSystems.Support;
+using WeaponCore.Data.Scripts.CoreSystems.Support.FireDistribution;
 
 namespace CoreSystems
 {
